move classification scripts to scripts
