+++ folding/utils/ops.py
@@ -2,7 +2,6 @@ import os
 import tqdm
 import random
 import subprocess
-import concurrent.futures
 import pickle as pkl
 from typing import List, Dict
 import requests
@@ -134,143 +133,3 @@ def is_pdb_complete(pdb_text: str) -> bool:
         if value in pdb_text_lower:
             return False
     return True
-
-
-def save_pkl(file, data):
-    with open(file, "wb") as f:
-        pkl.dump(data, f)
-
-
-def classify_pdb_batch(data: Dict, verbose=False):
-    """Downloads PDB files from a batch of PDB IDs and classifies them into complete, incomplete, and not downloadable lists. Saves the results to pickle files.
-
-    Args:
-        data (defaultdict[List]): A batch of PDB IDs, as returned by scripts/gather_pdbs.py.
-        verbose (bool, optional): If True, print the time required by the analysis and the percentages + frequencies of each list. Defaults to False.
-
-    Returns:
-        None
-    """
-    number_of_pdb_ids = sum([len(v) for v in data.values()])
-
-    complete = []
-    incomplete = []
-    not_downloadable = []
-    count = 0
-    complete_file = "scripts/pdb_ids_complete.pkl"
-    incomplete_file = "scripts/pdb_ids_incomplete.pkl"
-    not_downloadable_file = "scripts/pdb_ids_not_downloadable.pkl"
-
-    for v in tqdm(data.values()):
-        for pdb_id in v:
-            count += 1
-
-            try:
-                result = download_pdb("./complete_pdbs/", pdb_id + ".pdb")
-                if result:  # PDB was correctly downloaded and is complete
-                    complete.append(pdb_id)
-                else:  # PDB was correctly downloaded but is incomplete
-                    incomplete.append(pdb_id)
-            except Exception:  # Unable to download PDB
-                not_downloadable.append(pdb_id)
-                continue
-
-            if count % 10 == 0:  # Saving progress for safety
-                save_pkl(file=complete_file, data=complete)
-                save_pkl(file=incomplete_file, data=incomplete)
-                save_pkl(file=not_downloadable_file, data=not_downloadable)
-
-    # Once the entire process is finished, we save all the data.
-    save_pkl(file=complete_file, data=complete)
-    save_pkl(file=incomplete_file, data=incomplete)
-    save_pkl(file=not_downloadable_file, data=not_downloadable)
-
-    if verbose:
-        complete_percentage = len(complete) / number_of_pdb_ids * 100
-        incomplete_percentage = len(incomplete) / number_of_pdb_ids * 100
-        not_downloadable_percentage = len(not_downloadable) / number_of_pdb_ids * 100
-
-        print("=====================================")
-        print("Analysis Summary:")
-        print(f"Total number of PDB IDs: {number_of_pdb_ids}")
-        print(f"Complete: {len(complete)} ({complete_percentage:.2f}%)")
-        print(f"Incomplete: {len(incomplete)} ({incomplete_percentage:.2f}%)")
-        print(
-            f"Not Downloadable: {len(not_downloadable)} ({not_downloadable_percentage:.2f}%)"
-        )
-
-    print(
-        f"Analysis done!\nFiles saved at {complete_file}, {incomplete_file}, and {not_downloadable_file}"
-    )
-    print("=====================================")
-
-
-def parallel_classify_pdb_batch(data, verbose=False):
-    """
-    Classifies PDB IDs in parallel and saves the results to pickle files.
-
-    Args:
-        data (dict): A dictionary containing PDB IDs to be classified into complete, incomplete and not_downloadable.
-        verbose (bool, optional): If True, prints an analysis summary. Defaults to False.
-
-    Returns:
-        None
-
-    Raises:
-        None
-
-    Example:
-        data = {
-            'group1': ['pdb1', 'pdb2', 'pdb3'],
-            'group2': ['pdb4', 'pdb5']
-        }
-        parallel_classify_pdb_batch(data, verbose=True)
-
-    """
-    number_of_pdb_ids = sum([len(v) for v in data.values()])
-    complete = []
-    incomplete = []
-    not_downloadable = []
-    complete_file = "scripts/pdb_ids_complete.pkl"
-    incomplete_file = "scripts/pdb_ids_incomplete.pkl"
-    not_downloadable_file = "scripts/pdb_ids_not_downloadable.pkl"
-
-    def process_pdb(pdb_id):
-        nonlocal complete, incomplete, not_downloadable
-        try:
-            result = download_pdb("./complete_pdbs/", pdb_id + ".pdb")
-            if result:
-                complete.append(pdb_id)
-            else:
-                incomplete.append(pdb_id)
-        except Exception:
-            not_downloadable.append(pdb_id)
-
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        futures = [
-            executor.submit(process_pdb, pdb_id) for v in data.values() for pdb_id in v
-        ]
-        concurrent.futures.wait(futures)
-
-    save_pkl(file=complete_file, data=complete)
-    save_pkl(file=incomplete_file, data=incomplete)
-    save_pkl(file=not_downloadable_file, data=not_downloadable)
-
-    if verbose:
-        complete_percentage = len(complete) / number_of_pdb_ids * 100
-        incomplete_percentage = len(incomplete) / number_of_pdb_ids * 100
-        not_downloadable_percentage = len(not_downloadable) / number_of_pdb_ids * 100
-
-        print("=====================================")
-        print("Analysis Summary:")
-        print(f"Total number of PDB IDs: {number_of_pdb_ids}")
-        print(f"Complete: {len(complete)} ({complete_percentage:.2f}%)")
-        print(f"Incomplete: {len(incomplete)} ({incomplete_percentage:.2f}%)")
-        print(
-            f"Not Downloadable: {len(not_downloadable)} ({not_downloadable_percentage:.2f}%)"
-        )
-
-    print(
-        f"Analysis done!\nFiles saved at {complete_file}, {incomplete_file}, and {not_downloadable_file}"
-    )
-    print("=====================================")
